Complete game() function and modify playRound()
Modify the playRound() function to return a value, '1' if you win,
'0' if you tie, and '-1' if you lose, and use this as a counter to
see if the player ends up winning against the computer. The game()
function keeps track of the wins and losses and runs the playRound()
function 5 times.